count consectutive, count subarray sum

+++ leetcode/DynamicProgramming.cpp
@@ -155,7 +155,7 @@ public:
 			min_from_right[i] = min(A[i], min_from_right[i + 1]);
 		}
 		int max_from_left = INT32_MIN;
-		for (int i = 0; i < A.size(); i++) {
+		for (size_t i = 0; i < A.size(); i++) {
 			max_from_left = max(max_from_left, A[i]);
 			if (max_from_left <= min_from_right[i + 1])
 				return i + 1;
